Moved TrainedModelAssignment utils and Metdata as well as ModelAliasMetadata
